test api call with cert
