+++ api/package.json
@@ -2,7 +2,7 @@
   "name": "",
   "version": "1.0.0",
   "description": "",
-  "main": "src/functions/*.js",
+  "main": "getUsers/*.js",
   "scripts": {
     "start": "func start",
     "test": "echo \"No tests yet...\""
