--- api/src/functions/message.js
@@ -1,9 +1,0 @@
-const { app } = require('@azure/functions');
-
-app.http('message', {
-    methods: ['GET', 'POST'],
-    authLevel: 'anonymous',
-    handler: async (request, context) => {
-        return { body: `Hello, from the API!` };
-    }
-});
